chore: move useIsMounted hook one level up
